figma-transformer: split effects contract
